% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-grob.R
\name{as.grob}
\alias{as.grob}
\alias{as.grob.expression}
\alias{as.grob.formula}
\alias{as.grob.function}
\alias{as.grob.ggplot}
\alias{as.grob.meme}
\alias{as.grob.trellis}
\alias{as.grob.eulergram}
\alias{as.grob.Heatmap}
\alias{as.grob.upset}
\alias{as.grob.pheatmap}
\alias{as.grob.magick-image}
\alias{as.grob.grob}
\title{as.grob}
\usage{
as.grob(plot)

\method{as.grob}{expression}(plot)

\method{as.grob}{formula}(plot)

\method{as.grob}{`function`}(plot)

\method{as.grob}{ggplot}(plot)

\method{as.grob}{meme}(plot)

\method{as.grob}{trellis}(plot)

\method{as.grob}{eulergram}(plot)

\method{as.grob}{Heatmap}(plot)

\method{as.grob}{upset}(plot)

\method{as.grob}{pheatmap}(plot)

\method{as.grob}{grob}(plot)
}
\arguments{
\item{plot}{base or grid plot, or graphic object generated by ggplot, lattice, etc.}
}
\value{
grob object
}
\description{
convert a plot to grob object
}
\examples{
as.grob(~barplot(1:10))
}
\author{
guangchuang yu
}
