% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-ggplot.R
\name{as.ggplot}
\alias{as.ggplot}
\title{as.ggplot}
\usage{
as.ggplot(plot, scale = 1, hjust = 0, vjust = 0, angle = 0, ...)
}
\arguments{
\item{plot}{base or grid plot, or graphic generated by ggplot, lattice, etc.}

\item{scale}{scale of the plot to be drawn}

\item{hjust}{horizontal adjustment}

\item{vjust}{vertical adjustment}

\item{angle}{angle to rotate plot}

\item{...}{additional parameters passed to as.grob}
}
\value{
ggplot object
}
\description{
convert plot to ggplot object
}
\examples{
as.ggplot(~barplot(1:10))
}
\author{
Guangchuang Yu
}
