% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-grob.R
\name{base2grob}
\alias{base2grob}
\title{base2grob}
\usage{
base2grob(x)
}
\arguments{
\item{x}{expression or formula of base plot function call, e.g. expression(pie(1:5)) or ~plot(1:10, 1:10);
or a function that plots to an R graphics device when called, e.g. function() plot(sqrt)}
}
\value{
grob object
}
\description{
convert base plot to grob object
}
\examples{
base2grob(~plot(rnorm(10)))
}
\author{
Guangchuang Yu
}
