% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotify-package.R
\docType{package}
\name{ggplotify-package}
\alias{ggplotify}
\alias{ggplotify-package}
\title{ggplotify: Convert Plot to 'grob' or 'ggplot' Object}
\description{
Convert plot function call (using expression or formula) to 'grob' or 'ggplot' object that compatible to the 'grid' and 'ggplot2' ecosystem. With this package, we are able to e.g. using 'cowplot' to align plots produced by 'base' graphics, 'ComplexHeatmap', 'eulerr', 'grid', 'lattice', 'magick', 'pheatmap', 'vcd' etc. by converting them to 'ggplot' objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GuangchuangYu/ggplotify}
  \item Report bugs at \url{https://github.com/GuangchuangYu/ggplotify/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

}
\keyword{internal}
