% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-grob.R
\name{grid2grob}
\alias{grid2grob}
\title{grid2grob}
\usage{
grid2grob(plot_call)
}
\arguments{
\item{plot_call}{plot function call}
}
\value{
grob object
}
\description{
convert grid plot to grob object
}
\author{
Guangchuang Yu
}
