% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{rotate}
\usage{
rotate(p, angle, scale = 0.9)
}
\arguments{
\item{p}{plot that supported by ggplotify to convert to ggplot object}

\item{angle}{user selected angle}

\item{scale}{amount to scale the plot (zoom in or zoom out)}
}
\value{
display figure with specific angle
}
\description{
rotate figure with specific angle
}
\author{
Guangchuang Yu
}
